Move `Element::basepoint()` to be a bare method
This makes it a little more ergonomic to use in the common case where the crate
name is used as a prefix.

Closes #18
